+++ .storybook/main.ts
@@ -37,3 +37,4 @@ const config: StorybookConfig = {
 }
 
 export default config
+