+++ src/components/AdvancedTable/Table.tsx
@@ -67,7 +67,8 @@ export function Table<TData>({
   tableSettings,
   onRowSelection,
   onColumnSizing,
-  onPaginationChange
+  onPaginationChange,
+  rowEventsProps
 }: TTableProps<TData>) {
   const [expandedRows, setExpandedRows] = useState<Set<string>>(new Set())
 
@@ -88,6 +89,7 @@ export function Table<TData>({
       id: ColumnId.Expand,
       accessorKey: 'expand',
       header: () => <span style={{ display: 'none' }}>Expand</span>,
+      enableSorting: false,
       size: 50,
       minSize: 50,
       maxSize: 50,
@@ -223,7 +225,10 @@ export function Table<TData>({
                         ))
                       : table.getRowModel().rows.map((row) => (
                           <React.Fragment key={row.id}>
-                            <tr className={classnames({ selected: row.getIsSelected() })}>
+                            <tr
+                              {...rowEventsProps}
+                              className={classnames({ selected: row.getIsSelected() })}
+                            >
                               {row.getVisibleCells().map((cell) => (
                                 <td
                                   className={classnames({