+++ src/components/Input/Input.tsx
@@ -205,7 +205,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputCustomProps>(
               size="small"
               {...leftIconProps}
               className={classNames('input__icon', 'input__icon--left', {
-                [leftIconProps.className || '']: !!leftIconProps.className
+                [`${leftIconProps.className}`]: !!leftIconProps.className
               })}
             />
           ) : null}
@@ -214,7 +214,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputCustomProps>(
               size="small"
               {...rightIconProps}
               className={classNames('input__icon', 'input__icon--right', {
-                [rightIconProps.className || '']: !!rightIconProps.className
+                [`${rightIconProps.className}`]: !!rightIconProps.className
               })}
             />
           ) : null}