+++ src/components/Menu/Menu.tsx
@@ -14,16 +14,18 @@ import classNames from 'classnames'
 
 export const Menu = (props: TMenuProps): ReactElement | null => {
   const {
+    additionalRef = null,
     menuItems = [],
     parentRef,
     onClose,
     isOpen,
     position = 'bottom-right',
     className = '',
+    withTitleOption = false,
     children
   } = props
   const [menuRef, setMenuRef] = useState<HTMLDivElement | null>(null)
-  useOnOutsideClick(menuRef, onClose, isOpen, useId())
+  useOnOutsideClick([menuRef, additionalRef], onClose, isOpen, useId())
   useHideOnScroll(onClose)
   const { left, top } = useGetElemPositions(parentRef)
   const { width, height } = useGetElemSizes(parentRef)
@@ -73,6 +75,7 @@ export const Menu = (props: TMenuProps): ReactElement | null => {
                 return (
                   <OptionItem
                     dataId={dataId}
+                    withTitleOption={withTitleOption}
                     disabled={disabled}
                     key={value}
                     data={{