+++ src/components/Select/Select/Select.tsx
@@ -35,6 +35,7 @@ export const Select = (props: TSingleSelectPropTypes): JSX.Element | null => {
     isValid,
     isSearchable = false,
     isDynamicSearchable = false,
+    trimSearchValue = false,
     disabled,
     dataId = '',
     placeHolder,
@@ -142,10 +143,14 @@ export const Select = (props: TSingleSelectPropTypes): JSX.Element | null => {
       return options
     }
 
+    const processedSearchValue = trimSearchValue
+      ? searchValue.trim().toLowerCase()
+      : searchValue.toLowerCase()
+
     return options.filter((dataItem) => {
       return (
         typeof dataItem.label === 'string' &&
-        dataItem.label.toLowerCase().includes(searchValue.toLowerCase())
+        dataItem.label.toLowerCase().includes(processedSearchValue)
       )
     })
   }, [searchValue, options])
@@ -216,6 +221,22 @@ export const Select = (props: TSingleSelectPropTypes): JSX.Element | null => {
 
   useHideOnResize(closeDropdown)
 
+  const rightIconOpenedProps = useMemo(() => {
+    return {
+      ...selectRightIconOpenedProps,
+      className: 'cursor-pointer pointer-events-unset',
+      onClick: () => setIsOpen(false)
+    }
+  }, [selectRightIconOpenedProps])
+
+  const rightIconProps = useMemo(() => {
+    return {
+      ...selectRightIconProps,
+      className: 'cursor-pointer pointer-events-unset',
+      onClick: () => setIsOpen(true)
+    }
+  }, [selectRightIconProps])
+
   return (
     <div
       data-id={`${dataId}-content`}
@@ -235,7 +256,7 @@ export const Select = (props: TSingleSelectPropTypes): JSX.Element | null => {
         onInput={onInputChange}
         required={isRequiredField}
         leftIconProps={leftIconProps}
-        rightIconProps={isOpen ? selectRightIconOpenedProps : selectRightIconProps}
+        rightIconProps={isOpen ? rightIconOpenedProps : rightIconProps}
         readonly={!searchValue && !isWithSearch}
         placeholder={placeHolder}
         value={searchValue || selectedOption?.label || ''}